refactor helm autodiscovery plugin
