+++ pkg/plugins/resources/dockerimage/main.go
@@ -5,23 +5,8 @@ import (
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/google/go-containerregistry/pkg/v1/remote"
 	"github.com/mitchellh/mapstructure"
-	"github.com/updatecli/updatecli/pkg/plugins/utils/docker"
 	"github.com/updatecli/updatecli/pkg/plugins/utils/version"
 )
-
-// Spec defines a specification for a "dockerimage" resource
-// parsed from an updatecli manifest file
-type Spec struct {
-	// [S][C][T] Architecture specifies the container image architecture such as `amd64`
-	Architecture string `yaml:",omitempty"`
-	// [S][C][T] Image specifies the container image such as `updatecli/updatecli`
-	Image string `yaml:",omitempty"`
-	// [C][T] Tag specifies the container image tag such as `latest`
-	Tag                   string `yaml:",omitempty"`
-	docker.InlineKeyChain `yaml:",inline" mapstructure:",squash"`
-	// [S] VersionFilter provides parameters to specify version pattern and its type like regex, semver, or just latest.
-	VersionFilter version.Filter `yaml:",omitempty"`
-}
 
 // DockerImage defines a resource of type "dockerimage"
 type DockerImage struct {
